requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.12.5
google-generativeai==0.8.1
neo4j==5.24.0
